requirements.txt
Flask==3.1.0
flask_cors==5.0.1
Werkzeug==3.1.3
python-dotenv==1.0.0
psycopg2-binary==2.9.10
